+++ tests/gen_reactnative/test_interface/expected/ios/Getter.m
@@ -16,11 +16,11 @@ RCT_EXTERN_METHOD(getBoolean
                  withResolver:(RCTPromiseResolveBlock)resolve
                  withRejecter:(RCTPromiseRejectBlock)reject)
 
-RCT_EXTERN_METHOD(getInt
+RCT_EXTERN_METHOD(getLong
                  withResolver:(RCTPromiseResolveBlock)resolve
                  withRejecter:(RCTPromiseRejectBlock)reject)
 
-RCT_EXTERN_METHOD(getFloat
+RCT_EXTERN_METHOD(getDouble
                  withResolver:(RCTPromiseResolveBlock)resolve
                  withRejecter:(RCTPromiseRejectBlock)reject)
